feat: makes staff editbale
