+++ src/components/layouts/About/OurTeam/Board.tsx
@@ -1,9 +1,30 @@
 import { Box, Grid, Heading, Flex, Text } from "@chakra-ui/react";
 import PeaceImage from "components/atoms/PeaceImage";
+import Editable from "components/organisms/Editable/Editable";
+import { combinedConfig } from "config/constants/editable-copy/combined";
+import { useCopyData } from "contexts/EditableCopyContext";
+import { combinedConstant } from "config/constants/editable-copy/combined";
 import React from "react";
-import { boards } from "../constants";
 
+export type BoardMember = {
+  image: string;
+  name: string;
+  post: string;
+};
+
+export type BoardMembers = {
+  board: BoardMember[];
+};
 const Board = () => {
+  const { aboutConfig } = combinedConfig;
+
+  const { data } = useCopyData();
+
+  const { boards } = {
+    ...combinedConstant.boardMembers,
+    ...data.boardMembers,
+  };
+
   return (
     <Box>
       <Box>
@@ -32,70 +53,76 @@ const Board = () => {
         </Heading>
       </Box>
 
-      <Grid
-        gridTemplateColumns="1fr 1fr 1fr 1fr"
-        gridGap="8px"
-        mb="100px"
-        display={{
-          sm: "none",
-          md: "none",
-          lg: "none",
-          xl: "grid",
-          "2xl": "grid",
-        }}
+      <Editable
+        defaultValues={boards}
+        config={aboutConfig.board.boardMembers}
+        page="boardMembers"
       >
-        {boards.map((board, index) => {
-          return (
-            <Box key={index} w={{ xl: "260px", "2xl": "260px" }}>
-              <PeaceImage
-                src={board.image}
-                alt="board"
-                w="100%"
-                h="250px"
-                borderRadius="5px"
-              />
+        <Grid
+          gridTemplateColumns="1fr 1fr 1fr 1fr"
+          gridGap="8px"
+          mb="100px"
+          display={{
+            sm: "none",
+            md: "none",
+            lg: "none",
+            xl: "grid",
+            "2xl": "grid",
+          }}
+        >
+          {boards.map((board, index) => {
+            return (
+              <Box key={index} w={{ xl: "260px", "2xl": "260px" }}>
+                <PeaceImage
+                  src={board.image}
+                  alt="board"
+                  w="100%"
+                  h="250px"
+                  borderRadius="5px"
+                />
 
-              <Flex
-                alignItems="center"
-                bg="rgba(255, 255, 255, 67%)"
-                p="10px 25px"
-                position="relative"
-                h="70px"
-                top="-70px"
-                borderRadius="0px 0px 5px 5px"
-              >
-                <Box
-                  bg="#021D37"
-                  w="3px"
-                  my="auto"
-                  h="80%"
-                  position="absolute"
-                  zIndex="1"
-                ></Box>
+                <Flex
+                  alignItems="center"
+                  bg="rgba(255, 255, 255, 67%)"
+                  p="10px 25px"
+                  position="relative"
+                  h="70px"
+                  top="-70px"
+                  borderRadius="0px 0px 5px 5px"
+                >
+                  <Box
+                    bg="#021D37"
+                    w="3px"
+                    my="auto"
+                    h="80%"
+                    position="absolute"
+                    zIndex="1"
+                  ></Box>
 
-                <Box ml="10px">
-                  <Heading
-                    fontFamily="Playfair Display"
-                    fontSize={{ xl: "18px", "2xl": "22px" }}
-                    fontWeight="700"
-                    color="#021D37"
-                  >
-                    {board.name}
-                  </Heading>
-                  <Text
-                    fontFamily="Manrope"
-                    fontSize="14px"
-                    fontWeight="400"
-                    color="rgba(2, 29, 55, 0.63);"
-                  >
-                    {board.post}
-                  </Text>
-                </Box>
-              </Flex>
-            </Box>
-          );
-        })}
-      </Grid>
+                  <Box ml="10px">
+                    <Heading
+                      fontFamily="Playfair Display"
+                      fontSize={{ xl: "18px", "2xl": "22px" }}
+                      fontWeight="700"
+                      color="#021D37"
+                    >
+                      {board.name}
+                    </Heading>
+                    <Text
+                      fontFamily="Manrope"
+                      fontSize="14px"
+                      fontWeight="400"
+                      color="rgba(2, 29, 55, 0.63);"
+                    >
+                      {board.post}
+                    </Text>
+                  </Box>
+                </Flex>
+              </Box>
+            );
+          })}
+        </Grid>
+      </Editable>
     </Box>
   );
 };
