+++ src/components/layouts/About/OurTeam/Management.tsx
@@ -3,11 +3,33 @@ import React from "react";
 import ResponsiveManagement from "./Media Query/Responsive Management";
 import ResponsiveBoard from "./Media Query/Responsive Board";
 import ResponsiveStaff from "./Media Query/Responsive Staff";
-import { managements } from "../constants";
 import MobileRedirect from "components/molecules/MobileRedirect";
 import PeaceImage from "components/atoms/PeaceImage";
+import Editable from "components/organisms/Editable/Editable";
+import { combinedConfig } from "config/constants/editable-copy/combined";
+import { useCopyData } from "contexts/EditableCopyContext";
+import { combinedConstant } from "config/constants/editable-copy/combined";
+
+export type ManagementType = {
+  image: string;
+  name: string;
+  post: string;
+};
+
+export type Managements = {
+  managements: ManagementType[];
+};
 
 const Management = () => {
+  const { aboutConfig } = combinedConfig;
+
+  const { data } = useCopyData();
+
+  const { management } = {
+    ...combinedConstant.managements,
+    ...data.managements,
+  };
+
   const redirects = [
     {
       name: "PROFILE",
@@ -45,70 +67,76 @@ const Management = () => {
         Management
       </Heading>
 
-      <Flex
-        display={{
-          sm: "none",
-          md: "none",
-          lg: "none",
-          xl: "flex",
-          "2xl": "flex",
-        }}
+      <Editable
+        defaultValues={management}
+        config={aboutConfig.managements.managements}
+        page="managements"
       >
-        {managements.map((management, index) => {
-          return (
-            <Box
-              key={index}
-              borderRadius="50px"
-              w={{ xl: "250px", "2xl": "250px" }}
-              mr="10px"
-            >
-              <PeaceImage
-                src={management.image}
-                alt="management"
-                h={250}
-                w="100%"
-              />
-
-              <Flex
-                alignItems="center"
-                bg="rgba(255, 255, 255, 67%)"
-                p="10px 25px"
-                position="relative"
-                top={{ xl: "-60px", "2xl": "-60px" }}
-                borderRadius="0px 0px 5px 5px"
+        <Flex
+          display={{
+            sm: "none",
+            md: "none",
+            lg: "none",
+            xl: "flex",
+            "2xl": "flex",
+          }}
+        >
+          {management.map((management, index) => {
+            return (
+              <Box
+                key={index}
+                borderRadius="50px"
+                w={{ xl: "250px", "2xl": "250px" }}
+                mr="10px"
               >
-                <Box
-                  bg="#021D37"
-                  w="3px"
-                  my="auto"
-                  h="80%"
-                  position="absolute"
-                  zIndex="1"
-                ></Box>
+                <PeaceImage
+                  src={management.image}
+                  alt="management"
+                  h={250}
+                  w="100%"
+                />
 
-                <Box ml="10px">
-                  <Heading
-                    fontFamily="Playfair Display"
-                    fontSize={{ xl: "18x", "2xl": "22px" }}
-                    fontWeight="700"
-                    color="#021D37"
-                  >
-                    {management.name}
-                  </Heading>
-                  <Text
-                    fontFamily="Manrope"
-                    fontSize="14px"
-                    fontWeight="400"
-                    color="rgba(2, 29, 55, 0.63);"
-                  >
-                    {management.post}
-                  </Text>
-                </Box>
-              </Flex>
-            </Box>
-          );
-        })}
-      </Flex>
+                <Flex
+                  alignItems="center"
+                  bg="rgba(255, 255, 255, 67%)"
+                  p="10px 25px"
+                  position="relative"
+                  top={{ xl: "-60px", "2xl": "-60px" }}
+                  borderRadius="0px 0px 5px 5px"
+                >
+                  <Box
+                    bg="#021D37"
+                    w="3px"
+                    my="auto"
+                    h="80%"
+                    position="absolute"
+                    zIndex="1"
+                  ></Box>
+
+                  <Box ml="10px">
+                    <Heading
+                      fontFamily="Playfair Display"
+                      fontSize={{ xl: "18x", "2xl": "22px" }}
+                      fontWeight="700"
+                      color="#021D37"
+                    >
+                      {management.name}
+                    </Heading>
+                    <Text
+                      fontFamily="Manrope"
+                      fontSize="14px"
+                      fontWeight="400"
+                      color="rgba(2, 29, 55, 0.63);"
+                    >
+                      {management.post}
+                    </Text>
+                  </Box>
+                </Flex>
+              </Box>
+            );
+          })}
+        </Flex>
+      </Editable>
 
       <Box
         display={{
